problem/serializers.py
from rest_framework import serializers
from .models import Problem


class ProblemSerializer(serializers.ModelSerializer):
    class Meta:
        model = Problem
        fields = '__all__'
        extra_kwargs = {'properties': {
            'required': False
        }}